+++ src/Properties/Amount.php
@@ -5,19 +5,32 @@ namespace GingerPluginSdk\Properties;
 use GingerPluginSdk\Bases\BaseField;
 use GingerPluginSdk\Exceptions\OutOfDiapasonException;
 use GingerPluginSdk\Helpers\FieldsValidatorTrait;
-use GingerPluginSdk\Helpers\HelperTrait;
 use GingerPluginSdk\Interfaces\ValidateFieldsInterface;
-use GingerPluginSdk\Interfaces\ValueInCentsInterface;
 
-class Amount extends BaseField implements ValueInCentsInterface, ValidateFieldsInterface
+/**
+ * Amount describes cost of item or order and multiplied by 100.
+ *
+ * Validation for Amount value is to be greater than 1.
+ */
+final class Amount extends BaseField implements ValidateFieldsInterface
 {
-    use HelperTrait, FieldsValidatorTrait;
+    use FieldsValidatorTrait;
 
-    public function __construct(float|int $value)
+    private string $property_name = 'amount';
+
+    /**
+     * User could initialize amount property by 2 ways :
+     * 1. Using value in cents, in this case user should only provide integer value into constructor.
+     * 2. Using RawCost property, in this case user should firstly initialize RawCost property and after provide it to a constructor.
+     * Different between 1 and 2 case in the fact, that RawCost property will be firstly converted to a value in cents.
+     *
+     * @param RawCost|int $value
+     */
+    public function __construct(RawCost|int $value)
     {
-        $this->propertyName = 'amount';
-        $this->set($this->calculateValueInCents($value));
-        parent::__construct('amount');
+        $value = (string)$value;
+        $this->set((int)$value);
+        parent::__construct($this->property_name);
     }
 
     /**
@@ -26,7 +39,7 @@ class Amount extends BaseField implements ValueInCentsInterface, ValidateFieldsI
     public function validate($value)
     {
         if ($value < 1) throw new OutOfDiapasonException(
-            propertyName: $this->propertyName,
+            propertyName: $this->property_name,
             value: $value,
             min: 1);
     }