+++ src/Properties/Currency.php
@@ -3,8 +3,6 @@
 namespace GingerPluginSdk\Properties;
 
 use GingerPluginSdk\Bases\BaseField;
-use GingerPluginSdk\Exceptions\OutOfEnumException;
-use GingerPluginSdk\Exceptions\OutOfPatternException;
 use GingerPluginSdk\Helpers\FieldsValidatorTrait;
 use GingerPluginSdk\Interfaces\ValidateFieldsInterface;