+++ src/Tests/EventsTest.php
@@ -19,8 +19,8 @@ class EventsTest extends \PHPUnit\Framework\TestCase
     public function test_add_event()
     {
         $event = new Event(
-            occurred: '2022-05-17T11:58:33.813534+00:00',
             event: 'new',
+            occurred: '2022-05-17T11:58:33.813534+00:00',
             source: '123',
             noticed: '2022-05-17T11:58:33.813534+00:00',
             id: '123'
@@ -36,8 +36,8 @@ class EventsTest extends \PHPUnit\Framework\TestCase
     public function test_remove_event()
     {
         $event = new Event(
-            occurred: '2022-05-17T11:58:33.813534+00:00',
             event: 'new',
+            occurred: '2022-05-17T11:58:33.813534+00:00',
             source: '123',
             noticed: '2022-05-17T11:58:33.813534+00:00',
             id: '123'